tests/requirements-chill.txt
pip-chill==1.0.1
pytest-console-scripts==1.1.0
